Fix build on non-Unix platforms
