+++ internal/executor/instance/persistentworker/isolation/parallels/vmkiller.go
@@ -1,0 +1,5 @@
+package parallels
+
+import "errors"
+
+var ErrVMKiller = errors.New("failed to kill running VM processes")
