+++ internal/executor/instance/persistentworker/isolation/parallels/vmkiller_unsupported.go
@@ -1,0 +1,12 @@
+// +build !darwin
+
+package parallels
+
+import (
+	"fmt"
+	"runtime"
+)
+
+func ensureNoVMsRunning() error {
+	return fmt.Errorf("%w: unsupported platform: %s", ErrVMKiller, runtime.GOOS)
+}
